Refactor TSRanger into separate Model, View, and Controller files
