+++ recovery.md
@@ -50,3 +50,13 @@
 
 **Next Steps:**
 - Execute Sprint 2 starting with Architect spec, then TUI core, completion integration, and execution bridge.
+
+## 2025-08-08 (later)
+
+**Summary:**
+- Followed README.md recovery guidance to proceed autonomously.
+- Added Sprint 2 task to refactor `TSRanger` into one class per TS file.
+- Completed the refactor by extracting `RangerModel`, `RangerView`, and `RangerController` into separate files and wiring via `TSRanger`.
+
+**Next Steps:**
+- Run tests and ensure no regressions. Address any issues if they arise.
